+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## v0.1.9
+
+- Fixed: Parsing `DateRange` error
+
+- Chore: Use timestamp as `DateRange` unit
+
 ## v0.1.8
 
 - Fixed: Do not import existing blogs repeatedly